[*] Move formatAmount.types inside component folder
