+++ src/UI/FormatAmount/FormatAmount.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import classNames from 'classnames';
 import globalStyles from 'assets/sass/main.scss';
 import { DECIMALS, DIGITS, MAINNET_EGLD_LABEL } from 'constants/index';
-import { FormatAmountPropsType } from 'UI/types/formatAmount.types';
 import { stringIsInteger } from 'utils/validation/stringIsInteger';
 import { formatAmount } from '../../utils/operations/formatAmount';
+import { FormatAmountPropsType } from './formatAmount.types';
 import styles from './formatAmountStyles.scss';
 
 const formatAmountInvalid = (props: FormatAmountPropsType) => {
