+++ src/UI/FormatAmount/index.ts
@@ -1,1 +1,2 @@
 export * from './FormatAmount';
+export * from './formatAmount.types';
