+++ src/UI/types/index.ts
@@ -1,2 +1,1 @@
 export * from './withClassname.types';
-export * from './formatAmount.types';
